# Fix: Reading optimization ( add custom timing for each mode )