@RaulTrombin 2023-10-18
@patrickelectric  can we bump release 1.2 after this?