@RaulTrombin 2023-10-20
> I'm a bit inclined that from is ti the best way to implement it, but as a trair. The information does not imply directly on a sleep time.

It seems reasonable, but do not seems so verbose,

I was thinking into something like:

///but it's not possible
pub Enum mode {
    SomeMode (mode_values{value:10, ready_check:...})
}

pub struct mode_values  {
    value:u8,
    ready_check:std::time::Duration
}
pub type mode_var = (u8, std::time::Duration);

;

@joaoantoniocardoso